v1

+++ easy-mybatis/src/main/java/com/joaby/easy/study/mybatis/mapper/TestMapper.java
@@ -7,5 +7,5 @@ import com.joaby.easy.study.mybatis.entity.Test;
  * @Date: Created in 2019/11/12 15:57
  */
 public interface TestMapper {
-    Test selectByPrimaryKey(Long id);
+    Test selectByPrimaryKey(Integer id);
 }
